SSN in Doctor, Employee, Person changed to long so the ssn can be carried

+++ src/shared/Employee.java
@@ -7,7 +7,7 @@ public abstract class Employee extends Person implements Serializable
   private DateTime startDate;
   private String education;
 
-  public Employee(String firstName, String middleName, String lastName, int ssn,
+  public Employee(String firstName, String middleName, String lastName, long ssn,
       DateTime dob, Address address,String contactFirstName, String contactMiddleName, String contactLastName, String contactPhoneNumber, DateTime startDate, String education)
   {
     super(firstName, middleName, lastName, ssn, dob, address,contactFirstName,  contactMiddleName, contactLastName,  contactPhoneNumber );
